Practica_01/requirements.txt
# requirements.txt

Pillow 
